* commit '0.15.0b2-14-g720d450': (133 commits)
  Changed solver from 'dense_cholesky' to 'cholesky' to eliminate deprecation warning.
  Added utility to skip tests if running on Travis
  MAINT: more explicit glob pattern in doc generation
  MAINT ensure that examples figures are displayed in the correct order
  #3356 - Added an exception raising when np.nan is passed into a HashingVectorizer.
  Corrected two typos in docstring.
  DOC updated installation documentation
  DOC: typo, envelop → envelope
  FIX #3372: unstable input check test for RANSACRegressor
  FIX multilabel deprecation warning in RidgeClassifierCV
  FIX assert_array_almost_equal for windows tests
  MAINT skip joblib multiprocessing tests on travis
  TST fix precision failure on windows
  DOC run optipng before uploading website
  Release 0.15.0b2
  MAINT remove residual sparsefuncs*.so when compiling
  DOC SVMs take string class labels as well as integers
  MAINT bump joblib to 0.8.2
  Convert windows_testing_downloader.ps1 from UTF-16le to UTF-8
  Metadata information for unpickling models in future versions
  ...

+++ sklearn/neighbors/classification.py
@@ -378,7 +378,7 @@ class RadiusNeighborsClassifier(NeighborsBase, RadiusNeighborsMixin,
                                  in zip(pred_labels[inliers], weights)],
                                 dtype=np.int)
 
-            mode = mode.ravel().astype(np.int)
+            mode = mode.ravel()
 
             y_pred[inliers, k] = classes_k.take(mode)
